refactor: create lrt component
