+++ app/components/Deposit/index.tsx
@@ -1,10 +1,11 @@
 "use client";
 
-import React, { useContext, useEffect } from 'react';
+import React, { useContext } from 'react';
+import LrtPopup from "./LrtPopup"
 import  { EthereumDataContext } from "@/app/context";
 import './styles.css';
 import { useState } from "react";
-import { Activity, Loading, Cross } from "../icons";
+import { Activity, Loading } from "../icons";
 import { DepositContent } from "./DepositContent";
 import { WithdrawContent } from "./WithdrawContent";
 import { ActivityContent } from "./ActivityContent";
@@ -31,21 +32,6 @@ const Deposit: React.FC<DepositProps> = ({ amountEther, setAmountEther }) => {
   const { pendingTransactions } = useTransaction();
   const { evmWallet } = useWallets();
 
-  const [lrtVisible, setLrtVisible] = useState<boolean>(false);
-
-  useEffect(() => {
-    const popupClosed = localStorage.getItem('lrtPopup');
-    console.log(popupClosed, "orrrcoo")
-    if (popupClosed === null) {
-      setLrtVisible(true);
-    }
-  }, []);
-
-  const closePopup = () => {
-    setLrtVisible(false);
-    localStorage.setItem('lrtPopup', 'true');
-  };
-
   return (
     <>
     <div className="deposit-container flex flex-col">
@@ -53,26 +39,7 @@ const Deposit: React.FC<DepositProps> = ({ amountEther, setAmountEther }) => {
           width: isModalOpen ? "0px" : "inherit", 
           paddingRight: activeTab === Tabs.Activity ? "8px" : "20px"
       }}>
-
-      { lrtVisible && 
-        <>
-          <div className="lrt-box flex flex-row justify-between items-center">
-            <div className="flex flex-col justify-between h-full">
-              <span className="text-[20px] font-semibold w-[247px] h-[52px] text-left line-[26px]" style={{lineHeight: "26px"}}>Deposit your LRTs to earn restaked yield</span>
-              <span className="text-[14px] font-medium text-left" style={{ color: "rgba(161, 254, 160, 1)"}}>Deposit Now</span> 
-            </div>
-            <div>
-              <img src="lrt.svg" alt="LRT image." className="h-[99px]" />
-            </div>
-            <div className="circle circle1"></div>
-            <div className="circle circle2"></div>
-            <div className="circle circle3"></div>
-            <div className="circle circle4"></div>
-            <div onClick={closePopup}><Cross crossClassName='lrt-cross' /></div>
-          </div>
-          <div className="dash-box" ></div>
-        </>
-      }
+        <LrtPopup />
 
         <div className="header-tabs" style={{
           marginRight: activeTab === Tabs.Activity ? "12px" : "0px"
